+++ src/Oro/Bundle/MessageQueueBundle/Test/Functional/MessageQueueExtension.php
@@ -3,13 +3,10 @@
 namespace Oro\Bundle\MessageQueueBundle\Test\Functional;
 
 use Oro\Bundle\MessageQueueBundle\Client\BufferedMessageProducer;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * It is expected that this trait will be used in classes that have "getContainer" static method.
  * E.g. classes derived from Oro\Bundle\TestFrameworkBundle\Test\WebTestCase.
- *
- * @method static ContainerInterface getContainer()
  */
 trait MessageQueueExtension
 {