+++ src/Oro/Bundle/UIBundle/Resources/public/js/app/modules/css-variable-module.js
@@ -1,3 +1,0 @@
-import cssVariablesManager from 'oroui/js/css-variables-manager';
-
-cssVariablesManager.initialize();